// src/generated/prisma/models/Customer.ts
export type CustomerUncheckedCreateInput = {
  id?: string
  address: string
  active?: boolean
  notes?: string | null
  userId: string
  createdAt?: Date | string
  updatedAt?: Date | string
  tickets?: Prisma.TicketUncheckedCreateNestedManyWithoutCustomerInput
}

export type CustomerUpdateInput = {
  id?: Prisma.StringFieldUpdateOperationsInput | string
  address?: Prisma.StringFieldUpdateOperationsInput | string
  active?: Prisma.BoolFieldUpdateOperationsInput | boolean
  notes?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  createdAt?: Prisma.DateTimeFieldUpdateOperationsInput | Date | string
  updatedAt?: Prisma.DateTimeFieldUpdateOperationsInput | Date | string
  tickets?: Prisma.TicketUpdateManyWithoutCustomerNestedInput
  user?: Prisma.UserUpdateOneRequiredWithoutCustomerNestedInput
}
// ...
export type NullableStringFieldUpdateOperationsInput = {
  set?: string | null
}